refactor mandate diff router
